_events_producers/requirements.txt
faker
sqlalchemy
pyodbc
requests
python-dotenv
azure-eventhub
sqlalchemy-access 
azure-identity
pymssql
python-tds
sqlalchemy-pytds